Use decimal instead of floats for currency

+++ src/records.rs
@@ -1,4 +1,5 @@
 use serde::Deserialize;
+use rust_decimal::Decimal;
 
 /// Defines a transaction type to the client's asset account
 #[derive(Deserialize, PartialEq, Debug)]
@@ -38,8 +39,5 @@ pub struct TransactionRecord {
     /// Transaction id, needed for disputes
     pub tx: TransactionId,
     /// Amount of money. Only available for deposits, withdrawal and chargebacks
-    // TODO: to change to some kind of fixed point decimal instead of floating point
-    // not really good idea adding currency in floating point numbers due to round/truncate etc.
-    // though for 4 digits precision should be fine?
-    pub amount: Option<f32>,
+    pub amount: Option<Decimal>,
 }
